requirements.txt
# Python 3.8.3
Bootstrap-Flask==1.3.2
click==7.1.2
dominate==2.5.1
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
PyYAML==5.3.1
visitor==0.1.3
Werkzeug==1.0.1
wheel==0.34.2
